requirements.txt
requests
snowflake-connector-python
python-dotenv
beautifulsoup4
lxml